refactor(index): extract Index.tsx logic to separate integration component
